Split CLI build graph test-command validation tests
