+++ tests/integration/cli_build.rs
@@ -10,5 +10,7 @@ mod emit_direct;
 mod frontend_json;
 #[path = "cli_build/graph_validation.rs"]
 mod graph_validation;
+#[path = "cli_build/graph_validation_test_command.rs"]
+mod graph_validation_test_command;
 #[path = "cli_build/legacy_graph_command.rs"]
 mod legacy_graph_command;
